+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# 1.7.3
+
+Released on Thursday, September 3 2026
+
+- Improved tracking of parse exceptions in the `HtmlParser` (#1315) @lahma
+- Fixed tokenizer buffer overrun on encoding change (#1290) @jafin
+- Fixed selector text dropping the nth-child of clause (#1292) @jafin
+- Fixed selector text dropping the case-insensitive (#1291) @jafin
+- Fixed selector specificity carrying between fields instead of saturating (#1300) @meziantou
+- Fixed condition check for disposed state in ReturnToPool (#1296) @meziantou
+- Fixed ArgumentOutOfRangeException CSS escapes (#1299) @meziantou
+- Fixed whitespace around An+B CSS selector notations (#1297)
+- Fixed unsupported pseudo-classes matching everything (#1295)
+- Fixed DomException for invalid foreign attribute names such as `xml:lang[` (#1294)
+- Fixed script data escaped state leaving escaped mode too early (#1298)
+- Fixed `CurrentScript` not indicating the currently executing script (#1308) @lahma
+- Fixed returned missing attribute-properties to use empty string instead of null (#1313) @lahma
+- Fixed removal of `data-*` attributes (#1310) @lahma
+- Added missing DomName/DomAccessor annotations (#1301) @meziantou
+- Added missing `DomLiterals` attribute to `AdjacentPosition` (#1311) @lahma
+
 # 1.7.2
 
 Released on Sunday, August 23 2026